(refactor) Move the rating options to a feedback specific helper
It has more general application than vacancies.
